+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/domain/Collector.java
@@ -38,6 +38,11 @@ public class Collector {
         this.person = person;
     }
 
+    public Collector(Student student, Person person) {
+        this.student = student;
+        this.person = person;
+    }
+
 
     public String getCollectedDate() {
         return collected_date;