+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/domain/Student.java
@@ -86,4 +86,8 @@ public class Student {
         return collectorList;
     }
 
+    /*public void addPerson(Person person) {
+        Collector collector = new Collector(this, person);
+        collectorList.add(collector);
+    }*/
 }